build-and-deploy.yml
Attempt towards a single CI environment for forking, building, deploying
that works equally for testing/forks as well as production publishing.

For me I get an auto-build when I push a branch called 'test' on my fork.
This will be the same for anyone forking the project to improve it, they
can manually run 'workflow_dispatch' or get an autobuild pushing to 'test'.

It is still possible to get a quick build like before using
workflow_dispatch selecting only BUILD_SINGLE_DOC and DEPLOY_ENV_TEST.

For production it should be possible to manually run 'workflow_dispatch'
from GH actions selecting the 3 options with "(live)" in the description.
BUILD_MULTI_DOC, FAILON_LINKCHECK, DEPLOY_ENV_LIVE.

The linkcheck is optional and options exist to allow bypass if there is
an temporary issue with a 3rd party website at the time of building.

Due to the way GH actions works, this commit needs to hit the master branch
before it shows up the controls for workflow_dispatch.

It is also advisable to setup an 'on schedule' to run once a month,
because GH pages expire after 3 months which removes all the
documentation.  So there is a step to control build settings for this
run but no entry at the top of file to setup a trigger.  Until the
project is happy it can run on auto-pilot.

The build process deals with:
  latest pseudo-tag for the latest version
  meta robots: to try to make test builds invisible to SEO
  linkcheck can be skipped, or enforced, or report only (non fatal)
  validating deployment to production have expected options