Update VSCodium.rst
So I dropped the search suggestion, and used your approach of giving them the path.  That is also better, because the new user is learning the directory structure. 

And I reviewed the whole document again, added a period here, and a sentence or two there. 

Thank you for giving me the chance to learn how these tools work.   Thank you for being so careful in how you do things, and in your blindingly fast response.  It makes SpinalHDL look really really good.