Replace the word "wire" by "signal"
This documentation use mostly "signal" so it's a bit confusing for a beginner  to have two names for the same thing. Also if you are new to HDL without a prior verilog experience (where the keyword is "wire"), wire make little sense, how can you have a 32 bit wire ("cable" would be the right term). A 32 bit signal on the other hand makes sense.